Moved search filter algorithm to helpers
